Drop dead sub-account login code + diagnostic logs
The Tencent and Volcengine console password-login flows were replaced
by the unified cookie-import + WebView login path in earlier commits
on this branch, so the RSA login pipeline, the per-provider session
managers, and the sub-account login form views are all dead code now.
Drop them along with the diagnostic SafeLog.warn breadcrumbs we left
across the cookie-save/load/resolve path while debugging the four
misc providers.

- Delete VolcengineLoginClient/AuthSign/RSAPublicKey + their tests.
- Delete TencentSessionManager + VolcengineSessionManager.
- Delete TencentSubAccountLoginRow + VolcengineSubAccountLoginRow
  views from the misc settings section (already unreachable from
  any case branch).
- Drop .mainAccountId / .subUsername / .subPassword from the
  MiscCredentialStore.Kind enum and the matching settings.json
  sensitive-key markers.
- Inline VolcengineResponseMetadata + VolcengineMetadataError into
  VolcengineQuotaAdapter (only remaining caller after the deletions).
