+++ Sources/VibeBarApp/Views/MiscProviderSettingsSection.swift
@@ -719,243 +719,6 @@ struct EnterpriseHostField: View {
     }
 }
 
-/// Sub-account password login form for Tencent Cloud Hunyuan.
-///
-/// Tencent's console exposes only a username/password sign-in path —
-/// no API key or OAuth — so the user creates a permission-restricted
-/// CAM sub-user, types `<sub-user>@<owner-uin-or-alias>` plus the
-/// password here, and Vibe Bar runs the login flow on demand.
-/// Username and password are stored in Keychain via
-/// `MiscCredentialStore`; `~/.vibebar/settings.json` never sees them.
-struct TencentSubAccountLoginRow: View {
-    @EnvironmentObject var environment: AppEnvironment
-    @EnvironmentObject var quotaService: QuotaService
-
-    @State private var loginDraft: String = ""
-    @State private var passwordDraft: String = ""
-    @State private var hasStored: Bool = false
-    @State private var status: String?
-    @State private var isSaving = false
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 4) {
-            HStack(spacing: 6) {
-                TextField("sub-user@<main-UID or alias>", text: $loginDraft)
-                    .textFieldStyle(.roundedBorder)
-                    .disableAutocorrection(true)
-                    .autocorrectionDisabled(true)
-                SecureField("Password", text: $passwordDraft)
-                    .textFieldStyle(.roundedBorder)
-                Button(isSaving ? "Saving…" : "Save", action: save)
-                    .disabled(loginDraft.isEmpty || passwordDraft.isEmpty || isSaving)
-                if hasStored {
-                    Button(role: .destructive, action: clear) {
-                        Image(systemName: "trash")
-                    }
-                    .help("Remove stored Tencent sub-account credentials")
-                }
-            }
-            HStack(spacing: 4) {
-                Image(systemName: hasStored ? "checkmark.circle.fill" : "info.circle")
-                    .foregroundStyle(hasStored ? Color.green : Color.secondary)
-                    .font(.caption)
-                Text(hasStored
-                     ? "Sub-account credentials saved in Keychain."
-                     : "Create a CAM sub-user with read-only Hunyuan access. Plain text over HTTPS — Tencent does not RSA-encrypt the password.")
-                    .font(.caption)
-                    .foregroundStyle(hasStored ? .secondary : .tertiary)
-                    .fixedSize(horizontal: false, vertical: true)
-            }
-            if let status {
-                Text(status)
-                    .font(.caption2)
-                    .foregroundStyle(.orange)
-            }
-        }
-        .onAppear { reloadStoredFlag() }
-    }
-
-    private func save() {
-        let trimmedLogin = loginDraft.trimmingCharacters(in: .whitespacesAndNewlines)
-        let trimmedPass = passwordDraft  // intentionally NOT trimmed — passwords may have padding
-        guard !trimmedLogin.isEmpty, !trimmedPass.isEmpty else { return }
-
-        // Accept either `<subUser>@<mainUID>`, `<subUser>@<alias>`, or just
-        // a sub-user. The owner part is required — we surface the error
-        // up-front rather than letting the login flow fail with a vague
-        // backend message.
-        guard let parsed = parseLogin(trimmedLogin) else {
-            status = "Format expected: sub-user@<main-account-UID or alias>."
-            return
-        }
-
-        isSaving = true
-        defer { isSaving = false }
-        let okMain = MiscCredentialStore.writeString(parsed.mainAccountId, tool: .tencentHunyuan, kind: .mainAccountId)
-        let okUser = MiscCredentialStore.writeString(parsed.subUsername, tool: .tencentHunyuan, kind: .subUsername)
-        let okPass = MiscCredentialStore.writeString(trimmedPass, tool: .tencentHunyuan, kind: .subPassword)
-        guard okMain, okUser, okPass else {
-            status = "Could not save credentials to Keychain."
-            return
-        }
-        status = nil
-        loginDraft = ""
-        passwordDraft = ""
-        hasStored = true
-
-        // Wipe any in-memory cookies from a prior account so the next
-        // refresh re-runs the login with the new credentials.
-        Task {
-            await TencentSessionManager.shared.wipeSession()
-            triggerRefresh()
-        }
-    }
-
-    private func clear() {
-        MiscCredentialStore.delete(tool: .tencentHunyuan, kind: .mainAccountId)
-        MiscCredentialStore.delete(tool: .tencentHunyuan, kind: .subUsername)
-        MiscCredentialStore.delete(tool: .tencentHunyuan, kind: .subPassword)
-        Task { await TencentSessionManager.shared.wipeSession() }
-        hasStored = false
-        status = "Credentials cleared."
-    }
-
-    private func reloadStoredFlag() {
-        hasStored = MiscCredentialStore.hasValue(tool: .tencentHunyuan, kind: .mainAccountId)
-            && MiscCredentialStore.hasValue(tool: .tencentHunyuan, kind: .subUsername)
-            && MiscCredentialStore.hasValue(tool: .tencentHunyuan, kind: .subPassword)
-    }
-
-    private func triggerRefresh() {
-        guard let account = environment.account(for: .tencentHunyuan) else { return }
-        Task { _ = await quotaService.refresh(account) }
-    }
-
-    private struct ParsedLogin {
-        let subUsername: String
-        let mainAccountId: String
-    }
-
-    private func parseLogin(_ raw: String) -> ParsedLogin? {
-        let parts = raw.split(separator: "@", maxSplits: 1)
-        guard parts.count == 2 else { return nil }
-        let user = parts[0].trimmingCharacters(in: .whitespacesAndNewlines)
-        let main = parts[1].trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !user.isEmpty, !main.isEmpty else { return nil }
-        return ParsedLogin(subUsername: user, mainAccountId: main)
-    }
-}
-
-/// Sub-account password login form for Volcengine / Doubao.
-///
-/// Volcengine's IAM uses a numeric main account UID and a sub-user
-/// name; the user types both plus the password. Vibe Bar runs the
-/// `encCerts` → RSA-encrypt → `mixtureLogin` flow on demand and
-/// caches the resulting session cookies in
-/// `VolcengineSessionManager`. Keychain stores the three Kind values
-/// (`mainAccountId`, `subUsername`, `subPassword`); settings.json
-/// never sees the password.
-struct VolcengineSubAccountLoginRow: View {
-    @EnvironmentObject var environment: AppEnvironment
-    @EnvironmentObject var quotaService: QuotaService
-
-    @State private var mainIdDraft: String = ""
-    @State private var usernameDraft: String = ""
-    @State private var passwordDraft: String = ""
-    @State private var hasStored: Bool = false
-    @State private var status: String?
-    @State private var isSaving = false
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 4) {
-            HStack(spacing: 6) {
-                TextField("Main account UID", text: $mainIdDraft)
-                    .textFieldStyle(.roundedBorder)
-                    .frame(maxWidth: 130)
-                    .disableAutocorrection(true)
-                TextField("Sub-user name", text: $usernameDraft)
-                    .textFieldStyle(.roundedBorder)
-                    .disableAutocorrection(true)
-                SecureField("Password", text: $passwordDraft)
-                    .textFieldStyle(.roundedBorder)
-                Button(isSaving ? "Saving…" : "Save", action: save)
-                    .disabled(mainIdDraft.isEmpty || usernameDraft.isEmpty || passwordDraft.isEmpty || isSaving)
-                if hasStored {
-                    Button(role: .destructive, action: clear) {
-                        Image(systemName: "trash")
-                    }
-                    .help("Remove stored Volcengine sub-account credentials")
-                }
-            }
-            HStack(spacing: 4) {
-                Image(systemName: hasStored ? "checkmark.circle.fill" : "info.circle")
-                    .foregroundStyle(hasStored ? Color.green : Color.secondary)
-                    .font(.caption)
-                Text(hasStored
-                     ? "Sub-account credentials saved in Keychain."
-                     : "Create a CAM sub-user with read-only Doubao access. Password is RSA-encrypted before transit; only the encrypted blob hits Volcengine's API.")
-                    .font(.caption)
-                    .foregroundStyle(hasStored ? .secondary : .tertiary)
-                    .fixedSize(horizontal: false, vertical: true)
-            }
-            if let status {
-                Text(status)
-                    .font(.caption2)
-                    .foregroundStyle(.orange)
-            }
-        }
-        .onAppear { reloadStoredFlag() }
-    }
-
-    private func save() {
-        let trimmedMainId = mainIdDraft.trimmingCharacters(in: .whitespacesAndNewlines)
-        let trimmedUser = usernameDraft.trimmingCharacters(in: .whitespacesAndNewlines)
-        let trimmedPass = passwordDraft  // intentionally not trimmed
-        guard !trimmedMainId.isEmpty, !trimmedUser.isEmpty, !trimmedPass.isEmpty else { return }
-        guard trimmedMainId.allSatisfy(\.isNumber) else {
-            status = "Main account UID must be all digits."
-            return
-        }
-
-        isSaving = true
-        defer { isSaving = false }
-        let okMain = MiscCredentialStore.writeString(trimmedMainId, tool: .volcengine, kind: .mainAccountId)
-        let okUser = MiscCredentialStore.writeString(trimmedUser, tool: .volcengine, kind: .subUsername)
-        let okPass = MiscCredentialStore.writeString(trimmedPass, tool: .volcengine, kind: .subPassword)
-        guard okMain, okUser, okPass else {
-            status = "Could not save credentials to Keychain."
-            return
-        }
-        status = nil
-        passwordDraft = ""
-        hasStored = true
-        Task {
-            await VolcengineSessionManager.shared.wipeSession()
-            triggerRefresh()
-        }
-    }
-
-    private func clear() {
-        MiscCredentialStore.delete(tool: .volcengine, kind: .mainAccountId)
-        MiscCredentialStore.delete(tool: .volcengine, kind: .subUsername)
-        MiscCredentialStore.delete(tool: .volcengine, kind: .subPassword)
-        Task { await VolcengineSessionManager.shared.wipeSession() }
-        hasStored = false
-        status = "Credentials cleared."
-    }
-
-    private func reloadStoredFlag() {
-        hasStored = MiscCredentialStore.hasValue(tool: .volcengine, kind: .mainAccountId)
-            && MiscCredentialStore.hasValue(tool: .volcengine, kind: .subUsername)
-            && MiscCredentialStore.hasValue(tool: .volcengine, kind: .subPassword)
-    }
-
-    private func triggerRefresh() {
-        guard let account = environment.account(for: .volcengine) else { return }
-        Task { _ = await quotaService.refresh(account) }
-    }
-}
-
 /// "Sign in via Web" affordance for cookie-based misc providers whose
 /// auto-import path is unreliable on the user's browser. Currently
 /// renders for any tool that `MiscWebLoginRegistry` knows how to drive.
